# docs/public/openapi.json
{
  "components": {
    "schemas": {
...
      "GatewayCatalogListResponse": {
        "description": "Paginated Merge-style model catalog response.",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GatewayCatalogModel"
            },
            "title": "Data",
            "type": "array"
          },
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "object": {
            "default": "list",
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "GatewayCatalogListResponse",
        "type": "object"
      },
      "GatewayCatalogModel": {
        "description": "Merge Gateway-style canonical model catalog object.",
        "properties": {
          "availability_status": {
            "default": "available",
            "title": "Availability Status",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "vendors": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GatewayVendorModelMetadata"
            },
            "title": "Vendors",
            "type": "object"
          }
        },
        "required": [
          "model",
          "provider",
          "display_name",
          "vendors"
        ],
        "title": "GatewayCatalogModel",
        "type": "object"
      },
...
      "GatewayVendorListResponse": {
        "description": "Paginated Merge-style vendor catalog response.",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GatewayVendorResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "object": {
            "default": "list",
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "GatewayVendorListResponse",
        "type": "object"
      },
      "GatewayVendorModelMetadata": {
        "description": "Execution metadata for a model served by one vendor.",
        "properties": {
          "availability_status": {
            "default": "available",
            "title": "Availability Status",
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/GatewayCatalogCapabilities"
          },
          "context_window": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          },
          "launch_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launch Date"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "pricing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GatewayCatalogPricing"
              },
              {
                "type": "null"
              }
            ]
          },
          "zero_data_retention": {
            "default": false,
            "title": "Zero Data Retention",
            "type": "boolean"
          }
        },
        "title": "GatewayVendorModelMetadata",
        "type": "object"
      },
      "GatewayVendorResponse": {
        "description": "Merge Gateway-style execution vendor object.",
        "properties": {
          "availability_status": {
            "default": "active",
            "title": "Availability Status",
            "type": "string"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "supports_byok": {
            "default": true,
            "title": "Supports Byok",
            "type": "boolean"
          },
          "supports_zdr": {
            "default": false,
            "title": "Supports Zdr",
            "type": "boolean"
          },
          "vendor": {
            "title": "Vendor",
            "type": "string"
          }
        },
        "required": [
          "vendor",
          "name",
          "models"
        ],
        "title": "GatewayVendorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelListResponse": {
        "description": "OpenAI-compatible model list response.",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelObject"
            },
            "title": "Data",
            "type": "array"
          },
          "object": {
            "default": "list",
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "ModelListResponse",
        "type": "object"
      },
      "ModelObject": {
        "description": "OpenAI-compatible model object.",
        "properties": {
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object": {
            "default": "model",
            "title": "Object",
            "type": "string"
          },
          "owned_by": {
            "title": "Owned By",
            "type": "string"
          },
          "pricing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPricingInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "created",
          "owned_by"
        ],
        "title": "ModelObject",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A clean FastAPI gateway for otari with API key management",
    "title": "otari-gateway",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models": {
      "get": {
        "description": "List all available models.\n\nReturns models auto-discovered from configured providers, enriched with\npricing data from the model_pricing table when available. Models that only\nexist in the pricing table are also included for backward compatibility.",
        "operationId": "list_models_v1_models_get",
        "parameters": [
          {
            "description": "Filter models by provider name",
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter models by provider name",
              "title": "Provider"
            }
          },
          {
            "description": "Filter gateway catalog models by execution vendor",
            "in": "query",
            "name": "vendor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter gateway catalog models by execution vendor",
              "title": "Vendor"
            }
          },
          {
            "description": "Fetch one gateway catalog model by provider/model id",
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fetch one gateway catalog model by provider/model id",
              "title": "Model"
            }
          },
          {
            "description": "Response format: 'openai' (default) or 'gateway'/'merge' catalog shape",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "openai",
              "description": "Response format: 'openai' (default) or 'gateway'/'merge' catalog shape",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "description": "Gateway catalog pagination cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Gateway catalog pagination cursor",
              "title": "Cursor"
            }
          },
          {
            "description": "Gateway catalog page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Gateway catalog page size",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ModelListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/GatewayCatalogListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/GatewayCatalogModel"
                    }
                  ],
                  "title": "Response List Models V1 Models Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/{model_id}": {
      "get": {
        "description": "Get details for a specific model.",
        "operationId": "get_model_v1_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelObject"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model",
        "tags": [
          "models"
        ]
      }
    },
...
    "/v1/vendors": {
      "get": {
        "description": "List Merge-style execution vendors and the models they can serve.",
        "operationId": "list_vendors_v1_vendors_get",
        "parameters": [
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor",
              "title": "Cursor"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayVendorListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Vendors",
        "tags": [
          "models",
          "vendors"
        ]
      }
    },
    "/v1/vendors/{vendor_id}": {
      "get": {
        "description": "Fetch one Merge-style execution vendor by id.",
        "operationId": "get_vendor_v1_vendors__vendor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "vendor_id",
            "required": true,
            "schema": {
              "title": "Vendor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayVendorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vendor",
        "tags": [
          "models",
          "vendors"
        ]
      }
    }
  }
}